+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - MariaDB: Hash password in privileges (bug #1166, regression from 4.7.2)
 - MariaDB, PostgreSQL: Use CREATE OR REPLACE FUNCTION if possible
 - PostgreSQL: Fix editing jsonb[] columns (bug #1300, regression from 5.4.2)
+- PostgreSQL 18: Support VIRTUAL generated columns (bug #1286)
 - Plugins: Method menuActions() (bug #1303)
 
 ## Adminer 5.4.3 (released 2026-07-09)